+++ src/modules/xhttp_prom/prom_metric.h
@@ -78,7 +78,7 @@ int prom_gauge_reset(str *s_name, str *l1, str *l2, str *l3);
 /**
  * @brief Add some positive amount to a counter.
  */
-int prom_counter_inc(str *s_name, int number, str *l1, str *l2, str *l3);
+int prom_counter_inc(str *s_name, unsigned long long number, str *l1, str *l2, str *l3);
 
 /**
  * @brief Set a value in a gauge.
